+++ example/src/SampleApp/ExampleCMSView.tsx
@@ -3,7 +3,7 @@ import { Box, Button } from "@material-ui/core";
 
 import {
     buildSchema,
-    useAuthContext,
+    useAuthController,
     useSideEntityController,
     useSnackbarController
 } from "@camberi/firecms";
@@ -21,7 +21,7 @@ export function ExampleCMSView() {
     const sideEntityController = useSideEntityController();
 
     // hook to do operations related to authentication
-    const authController = useAuthContext();
+    const authController = useAuthController();
 
     const customProductSchema = buildSchema({
         name: "Custom product",